metro_backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dlib==19.24.2
dnspython==2.6.1
exceptiongroup==1.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.110.0
h11==0.14.0
idna==3.6
motor==3.3.2
numpy==1.26.4
pillow==10.2.0
pydantic==2.6.3
pydantic_core==2.16.3
pymongo==4.6.2
requests==2.31.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
